Made point hash platform-agnostic
Hello from my 32-bit desktop.
At any rate, this now uses sizeof(size_t) to determine the shift amount.
If you're on a 64-bit machine, it does the same as it did before.
If you're on a 32-bit machine, you can actually compile it now!
Also switched to XOR instead of OR so that if there's overlap,
we get more varied results.