Cleaned up printing
Printing is now handled using PeriodicRunner,
a class I threw together during a hackathon.
Essentially it only runs what you pass it if it's been
a given amount of time since the last time it was called.
It uses a mutex to make it thread-safe, which is unnecessary here
but may be userful later and does not seem to impact performance.